update test imports after refactor
