+++ test/browser-no-default.test.ts
@@ -3,8 +3,8 @@ import { existsSync } from 'fs'
 import { mkdir, rm } from 'fs/promises'
 import { fileURLToPath } from 'url'
 import { install } from '@puppeteer/browsers'
-import { makeLogger } from './utils'
-import { findOrInstallBrowser } from '@/utils'
+import { makeLogger } from './utils/index.js'
+import { findOrInstallBrowser } from '@/browser.js'
 
 vi.mock('puppeteer', async (originalImport) => {
     const cacheDir = fileURLToPath(new URL('./fixtures/.cache/browser-no-default/', import.meta.url))
