+++ test/index.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from 'vitest'
 
 test('default exports function', async () => {
-    const module = await import('@/.')
+    const module = await import('@/index.js')
     expect(Object.keys(module)).toStrictEqual(['default'])
     expect(module.default).toBeTypeOf('function')
 })
