+++ test/integration.test.ts
@@ -1,7 +1,7 @@
 import { beforeAll, describe, expect, test } from 'vitest'
-import pdf from '@/.'
+import pdf from '@/index.js'
 import { AstroConfig, AstroIntegration } from 'astro'
-import { Logger, makeLogger, parsePdf } from './utils'
+import { Logger, makeLogger, parsePdf } from './utils/index.js'
 import { cp, lstat, mkdir, rm } from 'fs/promises'
 import { fileURLToPath } from 'url'
 import { existsSync } from 'fs'
