+++ test/load-page.test.ts
@@ -1,8 +1,8 @@
 import { describe, test, expect, beforeAll, afterAll } from 'vitest'
-import { start } from './utils/server'
+import { start } from './utils/server.js'
 import { Browser, launch, Page } from 'puppeteer'
 import { Server } from 'http'
-import { loadPage, PageError } from '@/page'
+import { loadPage, PageError } from '@/page.js'
 
 describe('load errors', () => {
     let server: Server
