+++ test/pathname-filepath.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from 'vitest'
 import { pathToFileURL } from 'url'
 import { resolve, sep } from 'path'
-import { filepathToPathname, pathnameToFilepath } from '@/utils'
+import { filepathToPathname, pathnameToFilepath } from '@/utils.js'
 
 function createTests(name: string, resolveRoot: (root: string) => string | URL, slash = '/') {
     const d = resolveRoot
