+++ test/tsconfig.json
@@ -1,13 +1,14 @@
 {
     "include": ["**/*"],
-    "exclude": ["fixtures/**/*"],
     "compilerOptions": {
         "strict": true,
-        "module": "es2020",
+        "skipLibCheck": true,
         "target": "es2022",
-        "moduleResolution": "node",
+        "module": "node16",
+        "moduleResolution": "node16",
         "paths": {
             "@/*": ["../src/*"]
-        }
+        },
+        "noEmit": true
     }
 }
